Testing parallelization strategies

+++ corpus_cleaner/cleaner.py
@@ -14,7 +14,7 @@ from typing import Iterable, List, Tuple
 from corpus_cleaner.components.cleaner_component import CleanerComponent
 from corpus_cleaner.document import Document
 from collections import OrderedDict
-from corpus_cleaner.par_utils import Pipeline, PipelineLogger
+from corpus_cleaner.par_utils import MappingPipeline, PipelineLogger, CollectingPipeline
 from typing import Generator
 from typing import cast
 import argparse
@@ -49,11 +49,17 @@ class Cleaner:
         self.logger = PipelineLogger(logger)
         self.args.logger = self.logger
         self.mappers = MAPPERS
+        self.tmp_dir = os.path.join(args.output_path, 'tmp')
+        os.makedirs(self.tmp_dir)
         if args.components is not None:
             self.mappers = []
             for comp in MAPPERS:
                 if comp.__name__ in args.components:
                     self.mappers.append(comp)
+        self.mappers = [lambda x: DataParserFactory.get_parser_mapper(x)] + self.mappers +\
+                       [lambda x: OutputFormatterFactory.get_output_formatter_mapper(
+                           args=None, output_format = 'onion',
+                           output_path=os.path.join(self.tmp_dir, str(os.getpid()) + '.onion'))]
         self.reducer = REDUCER
         if args.components is not None:
             self.reducer = None
@@ -93,6 +99,11 @@ class Cleaner:
         parser = DataParserFactory.get_parser(self.args)
         return parser.parse()
 
+    def _get_paths(self) -> List[Tuple[int, str]]:
+        # self.logger.info('Parsing...')
+        parser = DataParserFactory.get_parser(self.args)
+        return parser.get_idx_relative_filepaths()
+
     def _create_pipeline_mappers(self) -> List[CleanerComponent]:
         return [component(self.args) for component in self.mappers]
 
@@ -104,7 +115,16 @@ class Cleaner:
         output_formatter = OutputFormatterFactory.get_output_formatter(self.args)
         output_formatter.apply(documents)
 
-    def clean(self):
+    def null_func(self, *args):
+        pass
+
+    def build_identity(self):
+        return [self.identity]
+
+    def identity(self, x):
+        return x
+
+    def clean_(self):
         if self.reducer is None:
             pipeline = Pipeline(streamers=[cast(Generator, iterable) for iterable in self._get_documents()],
                                 mappers_factory=self._create_pipeline_mappers,
@@ -133,6 +153,41 @@ class Cleaner:
                                 log_every_iter=self.args.log_every_iter)
 
             pipeline.run()
+
+            # remove Onion temporary files
+            os.remove(self.reducer.onion_input_file)
+            os.remove(self.reducer.onion_output_file)
+
+
+    def clean(self):
+        if self.reducer is None:
+            raise NotImplementedError()
+        else:
+
+            self.reducer = self.reducer(self.args)
+            pipeline = MappingPipeline(streams=self._get_paths(),
+                                       mappers_factory=self._create_pipeline_mappers,
+                                       parallel=self.args.parallel,
+                                       logger=self.logger if self.args.log_every_iter != -1 else None,
+                                       log_every_iter=self.args.log_every_iter)
+            pipeline.run()
+
+            self.reducer.reduce()
+
+            self._output(self.reducer.get_documents()[0])
+
+            '''
+            pipeline = CollectingPipeline(streamers=[cast(Generator, iterable) for iterable in self.reducer.get_documents()],
+                                mappers_factory=self.build_identity,#self._create_pipeline_postmappers if len(POSTMAPPERS) > 1 else \
+                                    #self.build_identity,
+                                output_reducer=self._output, batch_size=self.args.batch_size,
+                                parallel=False, parallel_streams=False,
+                                logger=self.logger if self.args.log_every_iter != -1 else None,
+                                log_every_iter=self.args.log_every_iter)
+
+            pipeline.run()
+            '''
+
 
             # remove Onion temporary files
             os.remove(self.reducer.onion_input_file)
